Phase 9b: migrate skin stream callbacks to std::ptrdiff_t offset; fix WDGL nativeRasterOffset
- Change readNativeSkin/writeNativeSkin/getSizeNativeSkin in xboxskin.cpp,
  wdgl.cpp, ps2skin.cpp, and their declarations in rwxbox.h, rwwdgl.h,
  rwps2plg.h from int32 to std::ptrdiff_t offset
- Remove static_cast<int32>(offset) from skin.cpp dispatch helpers now
  that platform functions accept std::ptrdiff_t directly
- Migrate wdgl::nativeRasterOffset from int32 to
  plugin::PluginOffset<Raster, GlRaster> inside #ifdef RW_OPENGL;
  update rwwdgl.h extern declaration accordingly
- Inline createNativeRaster/destroyNativeRaster/copyNativeRaster as
  lambda bodies in registerNativeRaster and use plugin::extensionPtr
  in Texture::upload/bind

+++ src/skin.cpp
@@ -127,11 +127,11 @@ readSkin(Stream *stream, int32 len, void *object, std::ptrdiff_t offset, int32)
 	if(geometry->instData){
 		// TODO: function pointers
 		if(geometry->instData->platform == PLATFORM_PS2)
-			return ps2::readNativeSkin(stream, len, object, static_cast<int32>(offset));
+			return ps2::readNativeSkin(stream, len, object, offset);
 		else if(geometry->instData->platform == PLATFORM_WDGL)
-			return wdgl::readNativeSkin(stream, len, object, static_cast<int32>(offset));
+			return wdgl::readNativeSkin(stream, len, object, offset);
 		else if(geometry->instData->platform == PLATFORM_XBOX)
-			return xbox::readNativeSkin(stream, len, object, static_cast<int32>(offset));
+			return xbox::readNativeSkin(stream, len, object, offset);
 		else{
 			assert(0 && "unsupported native skin platform");
 			return nil;
@@ -186,11 +186,11 @@ writeSkin(Stream *stream, int32 len, void *object, std::ptrdiff_t offset, int32)
 
 	if(geometry->instData){
 		if(geometry->instData->platform == PLATFORM_PS2)
-			return ps2::writeNativeSkin(stream, len, object, static_cast<int32>(offset));
+			return ps2::writeNativeSkin(stream, len, object, offset);
 		else if(geometry->instData->platform == PLATFORM_WDGL)
-			return wdgl::writeNativeSkin(stream, len, object, static_cast<int32>(offset));
+			return wdgl::writeNativeSkin(stream, len, object, offset);
 		else if(geometry->instData->platform == PLATFORM_XBOX)
-			return xbox::writeNativeSkin(stream, len, object, static_cast<int32>(offset));
+			return xbox::writeNativeSkin(stream, len, object, offset);
 		else{
 			assert(0 && "unsupported native skin platform");
 			return nil;
@@ -232,11 +232,11 @@ getSizeSkin(void *object, std::ptrdiff_t offset, int32)
 
 	if(geometry->instData){
 		if(geometry->instData->platform == PLATFORM_PS2)
-			return ps2::getSizeNativeSkin(object, static_cast<int32>(offset));
+			return ps2::getSizeNativeSkin(object, offset);
 		if(geometry->instData->platform == PLATFORM_WDGL)
-			return wdgl::getSizeNativeSkin(object, static_cast<int32>(offset));
+			return wdgl::getSizeNativeSkin(object, offset);
 		if(geometry->instData->platform == PLATFORM_XBOX)
-			return xbox::getSizeNativeSkin(object, static_cast<int32>(offset));
+			return xbox::getSizeNativeSkin(object, offset);
 		if(geometry->instData->platform == PLATFORM_D3D8)
 			return -1;
 		if(geometry->instData->platform == PLATFORM_D3D9)
